:construction_worker: Move StepBox to its own component
