+++ app/routes/__layout/dette-kan-vi/workshops.smidighelsesjekk.tsx
@@ -8,9 +8,9 @@ import { json } from "@remix-run/server-runtime";
 
 import { Button } from "~/components/button";
 import { CapraImage } from "~/components/capra-image";
-import { Card } from "~/components/card";
 import { Pattern } from "~/components/pattern/pattern";
 import { Section } from "~/components/section";
+import { StepBox } from "~/components/step-box";
 import { TitleAndText } from "~/components/title-and-text";
 import { useHydrated } from "~/hooks/use-hydrated";
 import {
@@ -260,32 +260,3 @@ const InformationCard = ({
     </div>
   );
 };
-
-interface StepBoxProps {
-  titleAs: "h2" | "h3" | "h4";
-  title: React.ReactNode;
-  description?: React.ReactNode;
-  items?: React.ReactNode[];
-}
-const StepBox = ({
-  titleAs: TitleComponent,
-  title,
-  description,
-  items,
-}: StepBoxProps) => (
-  <Card as="li">
-    <article>
-      <TitleComponent className="font-bold uppercase text-main">
-        {title}
-      </TitleComponent>
-      <div>
-        <p>{description}</p>
-        <ul style={{ paddingLeft: 28, listStyle: "disc" }}>
-          {items?.map((item, index) => (
-            <li key={index}>{item}</li>
-          ))}
-        </ul>
-      </div>
-    </article>
-  </Card>
-);
